Gemini/3.Q&A_with_gemini/requirements.txt
streamlit 
ipykernel 
google-generativeai
python-dotenv
